% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_DevelTools.R
\name{add_dependencies_to_DESCRIPTION}
\alias{add_dependencies_to_DESCRIPTION}
\title{find and add all package dependencies to DESCRIPTION file}
\usage{
add_dependencies_to_DESCRIPTION(
  exact_version = "mongolite",
  ignore_folder = "devel",
  ignore_package = "RJobTissueArea"
)
}
\arguments{
\item{exact_version}{character vector of package names where exact version is obligatory}

\item{ignore_folder}{character vector of folder names in pkg directory which are build-ignored}

\item{ignore_package}{character vector of package names to remove from final list}
}
\value{
the DESCRIPTION file is overwritten
}
\description{
find and add all package dependencies to DESCRIPTION file
}
\examples{
add_dependencies_to_DESCRIPTION(
exact_version = c("mongolite","locfit","EBImage"),
ignore_folder <- c("devel"),
ignore_package <- "RJobTissueArea"
 )
}
\keyword{internal}
