% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImageProcessing.R
\name{attenuate_FLpeakValues}
\alias{attenuate_FLpeakValues}
\title{sets high FL values to a certain quantile FL value of a given attenuation quantile}
\usage{
attenuate_FLpeakValues(data_sum, attenuation = 0.01)
}
\arguments{
\item{attenuation}{}
}
\description{
attentuation defines the percentage of high FL values to be set to a lower cutoff value
}
\examples{
\donttest{

group_ID <- "P1761451"
output_dir <- "data_output"
RJobTissueArea:::create_working_directory(output_dir)
chip_IDs <- find_valid_group_chip_IDs(group_ID)
ScanHistory <- create_ScanHistory_extended(chip_IDs,
output_dir,
result_ID = group_ID)
image_groups <- create_hdr_image_groups(ScanHistory)
image_group_list<-image_groups$data[[1]]
data_sum <- calculate_data_sum(image_group_list)
data_att <- attenuate_FLpeakValues(data_sum,0.1)
m.data <- convert_image_vector_to_matrix(data_att)
grey_values <- create_pixmap_greyvalues(m.data,c(1,1))
grey_values\%>\%
imager::as.cimg()\%>\%
plot(main = "original dataSum")
}
}
