% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImageProcessing.R
\name{calculate_n_TissuePixel}
\alias{calculate_n_TissuePixel}
\title{calculate_n_TissuePixel}
\usage{
calculate_n_TissuePixel(pixelset)
}
\arguments{
\item{pixelset}{}
}
\description{
calculate_n_TissuePixel
}
\keyword{internal}
