% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImageProcessing.R
\name{calculate_perc_TissueArea}
\alias{calculate_perc_TissueArea}
\title{calculate_perc_TissueArea}
\usage{
calculate_perc_TissueArea(pixelset)
}
\arguments{
\item{pixelset}{}
}
\value{

}
\description{
calculate_perc_TissueArea
}
\keyword{internal}
