% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{check.if.valid}
\alias{check.if.valid}
\title{checks and rapairs the structure of FL.csv files}
\usage{
check.if.valid(input.data, chip.IDs, jobtype)
}
\arguments{
\item{input.data}{list of input dataframes}

\item{chip.IDs}{character of the chip_ID}
}
\description{
checks and rapairs the structure of FL.csv files
}
\concept{RJob mainscript (ZKWapp)}
\keyword{internal}
