% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_errorHandling.R
\name{check_EDLs_strings}
\alias{check_EDLs_strings}
\title{checks if EDL's contain readable XML-strings}
\usage{
check_EDLs_strings(EDLs, IDs = NULL, print = FALSE)
}
\arguments{
\item{EDLs}{character vector of XML-strings}

\item{IDs}{optional character vector of length(EDLs), containing IDs}

\item{print}{logical, error text should be printed}
}
\value{
EDLs character vector completed with error messages if neccessary
}
\description{
replaces EDLs by error messages if not a character or not readable
}
\examples{
testEDL <- c("<foo><bar /></foo>",
"<ObjRef
UID=\"D1891203\"
Type=\"Gate\"/><ObjRef
UID=\"D1614318\"
Type=\"Gate\"/><ObjRef
UID=\"D973219\"
Type=\"Gate\"/>",
"lalala",NA,NULL)

testEDL\%>\%check_EDLs_strings()

testEDL\%>\%check_EDLs_strings(IDs = c("ID1","ID2","ID3","ID4"))

testEDL\%>\%check_EDLs_strings(IDs = c("ID1"))

}
\seealso{
Other errorHandling: 
\code{\link{check_mappingResultLists}()},
\code{\link{detect_error_inCharacterVector}()},
\code{\link{get_errorDF_queryResults}()},
\code{\link{is_error_in_EDLs}()},
\code{\link{is_error_in_mappingResult}()},
\code{\link{print_queryResult_error}()},
\code{\link{split_errorText_toDF}()}
}
\concept{errorHandling}
\keyword{internal}
