% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{check_if_gatefile_exist}
\alias{check_if_gatefile_exist}
\title{Title}
\usage{
check_if_gatefile_exist(chip_path, segment_ID, gate_ID)
}
\arguments{
\item{gate_ID}{}
}
\value{

}
\description{
Title
}
