% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{check_imported_file_lists}
\alias{check_imported_file_lists}
\title{Check whether input files are valid}
\usage{
check_imported_file_lists(fl_file_names, gatenames)
}
\arguments{
\item{fl_file_names}{List of FL file names}

\item{gatenames}{List of gate file names}

\item{filesToProcess}{List of \code{.csv} files with fluorescence values,
tidied up by \code{\link{tidy_up_XML_file_content}}}

\item{gatesToProcess}{List of \code{.csv} files with cells contained with
gates, tidied up by \code{\link{tidy_up_XML_file_content}}}
}
\value{
List with \code{filesToProcess} without empty or not existing files,
and flag whether fatal errors occurred
}
\description{
This function checks whether at least one fluorescence file and gate files
are present. Empty files are removed.
}
\keyword{Check}
\keyword{file}
\keyword{imported}
\keyword{internal}
\keyword{lists}
\keyword{the}
