% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{check_input_files}
\alias{check_input_files}
\title{Title}
\usage{
check_input_files(filesToProcess, gatesToProcess)
}
\arguments{
\item{gatesToProcess}{}
}
\value{

}
\description{
Title
}
