% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_errorHandling.R
\name{check_mappingResultLists}
\alias{check_mappingResultLists}
\title{checks resultList and adds try-error text}
\usage{
check_mappingResultLists(
  result,
  IDs = NULL,
  error_text = "no_MethodHistory in EDLchannel"
)
}
\arguments{
\item{result}{result list gained withing \code{~try(...,silent=TRUE)}}

\item{IDs}{character vector of IDs}

\item{error_text}{character of a suffix to add}
}
\value{
resultList with errorText
}
\description{
takes a result list and check if each element contains a try-error output text
replaces try-error output by an error text including ID of the element and a optional note
}
\examples{
\donttest{

chip_ID <- c("M911981","test","M911981")

query_result <- query_UID_limslager(chip_ID)
query_error <- query_result$error_message

EDLs <- get_EDLs_of_queryResult(query_result)
EDLs_error <- is_error_in_EDLs(EDLs)

nodeset <- get_nodesets_of_EDLs(EDLs,chip_ID)
nodeset_error <- nodeset$error

MH <- extract_MethodHistory(nodeset)
MH <-  MH\%>\%check_mappingResultLists()
MH_error <- MH\%>\%is_error_in_mappingResult()

}
}
\seealso{
Other errorHandling: 
\code{\link{check_EDLs_strings}()},
\code{\link{detect_error_inCharacterVector}()},
\code{\link{get_errorDF_queryResults}()},
\code{\link{is_error_in_EDLs}()},
\code{\link{is_error_in_mappingResult}()},
\code{\link{print_queryResult_error}()},
\code{\link{split_errorText_toDF}()}
}
\concept{errorHandling}
\keyword{internal}
