% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chip_IDs}
\alias{chip_IDs}
\title{chip_IDs of a chipgroup}
\format{
a character vector of length 20
}
\source{
returned by package function: find_valid_group_chip_IDs(group_ID = "P1761451")
}
\usage{
data(chip_IDs)
}
\description{
A vector containing those chip_IDs for which data could be found.
}
\details{
\describe{
\item{chip_IDs}{IDs of chips}
}
}
\keyword{data}
