% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{collect_segments_metadata}
\alias{collect_segments_metadata}
\title{Title}
\usage{
collect_segments_metadata(chip_path)
}
\arguments{
\item{chip_path}{path to chip folder on image server}
}
\value{
df
}
\description{
description
}
\details{
details
}
\examples{
chip_path <- c("\\\\\\\\intern.chipcytometry.com\\\\imagedata\\\\leipzig_volume0\\\\M986054")
collect_segments_metadata(chip_path)
}
\keyword{internal}
