% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImportBlobFiles.R
\name{convert_binsize_from_encoding}
\alias{convert_binsize_from_encoding}
\title{convert_binsize_from_encoding}
\usage{
convert_binsize_from_encoding(encoding)
}
\arguments{
\item{encoding}{}
}
\description{
convert_binsize_from_encoding
}
\keyword{internal}
