% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImageProcessing.R
\name{convert_image_vector_to_matrix}
\alias{convert_image_vector_to_matrix}
\title{Title}
\usage{
convert_image_vector_to_matrix(data_sum)
}
\arguments{
\item{data_sum}{}
}
\value{

}
\description{
Title
}
