% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{copy_file_positionCSV}
\alias{copy_file_positionCSV}
\title{copy_file_positionCSV()}
\usage{
copy_file_positionCSV(xml_dir, chip_ID, output_dir)
}
\arguments{
\item{xml_dir}{external pointer to the content of the RService-XML-file}

\item{chip_ID}{character of the chip_ID}

\item{output_dir}{character of path to the output}
}
\description{
copys the position.csv into output_dir
}
\seealso{
Other RJob mainscript: 
\code{\link{create_XML_filename}()},
\code{\link{find_RService_XML_on_imageserver}()},
\code{\link{load_mainscript_output}()}
}
\concept{RJob mainscript}
\keyword{internal}
