% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImageProcessing.R
\name{create_export_data_sum}
\alias{create_export_data_sum}
\title{create_export_data_sum}
\usage{
create_export_data_sum(image_group_list, group_ID, output_dir)
}
\arguments{
\item{output_dir}{}
}
\value{

}
\description{
create_export_data_sum
}
\examples{
if(FALSE){

data_sum_resultion <- create_export_data_sum(image_groups$data[[1]],
image_groups$group_ID[1],
output_dir)

}
}
