% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{create_gatecsv_filepath}
\alias{create_gatecsv_filepath}
\title{Title}
\usage{
create_gatecsv_filepath(chip_path, segment_ID, gate_ID)
}
\arguments{
\item{gate_ID}{}
}
\value{

}
\description{
Title
}
