% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImportBlobFiles.R
\name{create_hdr_filepath}
\alias{create_hdr_filepath}
\title{create_hdr_filepath}
\usage{
create_hdr_filepath(chip_path, scan_ID, pos_ID)
}
\arguments{
\item{pos_ID}{}
}
\description{
create_hdr_filepath
}
\keyword{internal}
