% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImageProcessing.R
\name{create_hdr_image_groups}
\alias{create_hdr_image_groups}
\title{create_hdr_image_groups}
\usage{
create_hdr_image_groups(ScanHistory)
}
\arguments{
\item{ScanHistory}{}
}
\value{

}
\description{
create_hdr_image_groups
}
