% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_DBrelated.R
\name{create_long_node_df_from_XML}
\alias{create_long_node_df_from_XML}
\title{create_long_node_df_from_XML}
\usage{
create_long_node_df_from_XML(XML)
}
\arguments{
\item{XML}{}
}
\description{
create_long_node_df_from_XML
}
\keyword{internal}
