% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_MetadataScanPosition.R
\name{create_MethodHistory_from_EDL}
\alias{create_MethodHistory_from_EDL}
\title{Title}
\usage{
create_MethodHistory_from_EDL(EDL)
}
\arguments{
\item{EDL}{}
}
\value{

}
\description{
Title
}
\examples{
data(EDL_chipIDs_limslager)
EDL <- EDL_chipIDs_limslager[1]
MH <- create_MethodHistory_from_EDL(EDL)
}
