% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_MetadataScanPosition.R
\name{create_MethodHistory_from_EDLs}
\alias{create_MethodHistory_from_EDLs}
\title{Title}
\usage{
create_MethodHistory_from_EDLs(EDLs)
}
\arguments{
\item{EDLs}{character vector containing EDL strings}
}
\description{
Title
}
\examples{
data("EDL_chipIDs_limslager")
EDLs <- EDL_chipIDs_limslager
MHs <- create_MethodHistory_from_EDLs(EDLs)
}
\keyword{internal}
