% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImageProcessing.R
\name{create_name_result_ID}
\alias{create_name_result_ID}
\title{create_name_result_ID}
\usage{
create_name_result_ID(chip_ID, pos_ID, sigma, threshold)
}
\arguments{
\item{threshold}{}
}
\value{

}
\description{
create_name_result_ID
}
