% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImageProcessing.R
\name{create_pixmap_greyvalues}
\alias{create_pixmap_greyvalues}
\title{create_pixmap_greyvalues}
\usage{
create_pixmap_greyvalues(m.data, cellres)
}
\arguments{
\item{cellres}{}
}
\value{

}
\description{
create_pixmap_greyvalues
}
