% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImageProcessing.R
\name{create_plot_directory}
\alias{create_plot_directory}
\title{create_plot_directory}
\usage{
create_plot_directory(output_dir, sigma, threshold, window, chip_ID, pos_ID)
}
\arguments{
\item{pos_ID}{}
}
\value{

}
\description{
create_plot_directory
}
