% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImportBlobFiles.R
\name{create_pos_foldername}
\alias{create_pos_foldername}
\title{create_pos_foldername}
\usage{
create_pos_foldername(imageServer_path, basePath, pos_ID)
}
\arguments{
\item{pos_ID}{}
}
\value{

}
\description{
create_pos_foldername
}
\keyword{internal}
