% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImportBlobFiles.R
\name{create_Pos_image_filepath}
\alias{create_Pos_image_filepath}
\title{create_Pos_image_filepath}
\usage{
create_Pos_image_filepath(ScanBasePath, PositionFolder, Type)
}
\arguments{
\item{Type}{}
}
\value{

}
\description{
create_Pos_image_filepath
}
