% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_Cutoff.R
\name{create_query_result_nodes}
\alias{create_query_result_nodes}
\title{creates xml-nodes of query result EDL}
\usage{
create_query_result_nodes(query_result)
}
\arguments{
\item{query_result}{}
}
\value{

}
\description{
creates xml-nodes of query result EDL
}
\keyword{internal}
