% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImageProcessing.R
\name{create_result_filepath}
\alias{create_result_filepath}
\title{Title}
\usage{
create_result_filepath(output_dir, name_string, result_ID, type)
}
\arguments{
\item{type}{}
}
\description{
Title
}
\keyword{internal}
