% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_MetadataScanPosition.R
\name{create_ScanHistory_extended}
\alias{create_ScanHistory_extended}
\title{create_ScanHistory_extended}
\usage{
create_ScanHistory_extended(chip_IDs, output_dir, result_ID)
}
\arguments{
\item{chip_IDs}{character of chip_IDs}

\item{output_dir}{character of output directory}

\item{result_ID}{character of added to filename containing the result df}
}
\value{
a dataframe and a csv file exported into output_dir
}
\description{
create_ScanHistory_extended
}
\examples{
\donttest{
data("chip_IDs")
output_dir <- tempdir()
result_ID <- "test"
create_ScanHistory_extended(chip_IDs,output_dir,result_ID)
}
}
