% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_MetadataScanPosition.R
\name{create_ScanHistory_from_MethodHistory}
\alias{create_ScanHistory_from_MethodHistory}
\title{create_ScanHistory_from_MethodHistory}
\usage{
create_ScanHistory_from_MethodHistory(MethodHistory)
}
\arguments{
\item{MethodHistory}{}
}
\value{

}
\description{
create_ScanHistory_from_MethodHistory
}
\keyword{internal}
