% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_MetadataScanPosition.R
\name{create_ScanHistory_of_chipIDs}
\alias{create_ScanHistory_of_chipIDs}
\title{create_ScanHistory_of_chipIDs}
\usage{
create_ScanHistory_of_chipIDs(MethodHistory)
}
\arguments{
\item{MethodHistory}{}
}
\value{

}
\description{
create_ScanHistory_of_chipIDs
}
