% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_MetadataScanPosition.R
\name{create_ScanHistory_of_MethodHistory}
\alias{create_ScanHistory_of_MethodHistory}
\title{create_ScanHistory_of_MethodHistory}
\usage{
create_ScanHistory_of_MethodHistory(MethodHistory)
}
\arguments{
\item{MethodHistory}{}
}
\value{

}
\description{
create_ScanHistory_of_MethodHistory
}
\examples{
data("EDL_chipIDs_limslager")
EDLs <- EDL_chipIDs_limslager
MHs <- create_MethodHistory_from_EDLs(EDLs)
SHs <- create_ScanHistory_of_MethodHistory(MHs)
}
\keyword{internal}
