% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{create_working_directory}
\alias{create_working_directory}
\title{Create working directory if necessary}
\usage{
create_working_directory(output_dir)
}
\arguments{
\item{output_dir}{Given working directory}
}
\value{
The (existing) output directory
}
\description{
This function checks whether the provided path to the working directory is
valid, otherwise a path is constructed based on the current working
directory. If necessary, the working directory is created within the file
system.
}
