% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{create_XML_filename}
\alias{create_XML_filename}
\title{creates the filename of on RService-XML file}
\usage{
create_XML_filename(RJob_ID)
}
\arguments{
\item{RJob_ID}{character}
}
\value{
string of the filename
}
\description{
creates the filename of on RService-XML file
}
\examples{
create_XML_filename("test")
create_XML_filename(NA)
create_XML_filename(NULL)
}
\seealso{
Other RJob mainscript: 
\code{\link{copy_file_positionCSV}()},
\code{\link{find_RService_XML_on_imageserver}()},
\code{\link{load_mainscript_output}()}
}
\concept{RJob mainscript}
\keyword{internal}
