% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_Cutoff.R
\name{create_xml_nodes}
\alias{create_xml_nodes}
\title{creates xml-nodes of EDL query result}
\usage{
create_xml_nodes(EDL)
}
\arguments{
\item{EDL}{}
}
\value{

}
\description{
creates xml-nodes of EDL query result
}
\keyword{internal}
