% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_errorHandling.R
\name{detect_error_inCharacterVector}
\alias{detect_error_inCharacterVector}
\title{detects given string in character vector}
\usage{
detect_error_inCharacterVector(error, error_string = c("error", "Error"))
}
\arguments{
\item{error}{character vector}

\item{error_string}{character containing the string to detect}
}
\value{
logical if error text could be detected
}
\description{
detects given string in character vector
}
\examples{

error <- c(try(ln(0),silent = TRUE), "error_", NA,"<xml>")
detect_error_inCharacterVector(error)

}
\seealso{
Other errorHandling: 
\code{\link{check_EDLs_strings}()},
\code{\link{check_mappingResultLists}()},
\code{\link{get_errorDF_queryResults}()},
\code{\link{is_error_in_EDLs}()},
\code{\link{is_error_in_mappingResult}()},
\code{\link{print_queryResult_error}()},
\code{\link{split_errorText_toDF}()}
}
\concept{errorHandling}
\keyword{internal}
