% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImageProcessing.R
\name{detect_tissueArea}
\alias{detect_tissueArea}
\title{detect_tissueArea}
\usage{
detect_tissueArea(m.data, cellres, sigma, threshold, window)
}
\arguments{
\item{window}{}
}
\description{
detect_tissueArea
}
\keyword{internal}
