% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_MetadataScanPosition.R
\name{determine_scan_position}
\alias{determine_scan_position}
\title{Title}
\usage{
determine_scan_position(MethodHistory)
}
\arguments{
\item{MethodHistory}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
