% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_DevelTools.R
\name{download_pkg_sources}
\alias{download_pkg_sources}
\title{Title}
\usage{
download_pkg_sources(Package, Version, lib_path = "devel/pkg_sources")
}
\arguments{
\item{lib_path}{}
}
\value{

}
\description{
Title
}
\examples{
if(FALSE){
dep_pkg <- get_package_dependencies()
Package <- dep_pkg$Package
Version <- dep_pkg$Version
download_pkg_sources(Package,Version)
}
}
\keyword{internal}
