% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{execute_mainscript_FDR}
\alias{execute_mainscript_FDR}
\title{Title}
\usage{
execute_mainscript_FDR(chip_ID, markernames, RJob_ID, jobtype)
}
\arguments{
\item{jobtype}{}
}
\value{

}
\description{
Title
}
