% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImportBlobFiles.R
\name{export_blob_parameter_of_image_filelist}
\alias{export_blob_parameter_of_image_filelist}
\title{export_blob_parameter_of_image_filelist}
\usage{
export_blob_parameter_of_image_filelist(image_files, output_dir, result_ID)
}
\arguments{
\item{result_ID}{}
}
\value{

}
\description{
export_blob_parameter_of_image_filelist
}
