% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{export_full_input}
\alias{export_full_input}
\title{Export entire data set}
\usage{
export_full_input(
  all_files,
  fl_files,
  gate_files,
  chip_ID,
  RJob_ID,
  output_dir
)
}
\arguments{
\item{all_files}{List of all relevant files with metadata as attributes}

\item{fl_files}{List of FL file content}

\item{gate_files}{List of gate file content}

\item{chip_ID}{Chip UId}

\item{RJob_ID}{R Job UID}

\item{output_dir}{Output directory}
}
\description{
Export entire data set
}
\keyword{internal}
