% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImageProcessing.R
\name{export_image_result_tiffs}
\alias{export_image_result_tiffs}
\title{export_image_result_tiffs}
\usage{
export_image_result_tiffs(
  image,
  image_blurred,
  image_binary,
  output_dir,
  chip_ID,
  pos_ID,
  sigma,
  threshold,
  ncols,
  nrows
)
}
\arguments{
\item{nrows}{}
}
\value{

}
\description{
export_image_result_tiffs
}
