% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImportBlobFiles.R
\name{export_list_all_image_files}
\alias{export_list_all_image_files}
\title{export_list_all_image_files}
\usage{
export_list_all_image_files(chip_IDs, result_ID, output_dir)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
export_list_all_image_files
}
