% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImageProcessing.R
\name{export_tissueAreaResults}
\alias{export_tissueAreaResults}
\title{Title}
\usage{
export_tissueAreaResults(resultList, output_dir, result_ID)
}
\arguments{
\item{result_ID}{}
}
\description{
Title
}
\keyword{internal}
