% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{extract_chip_IDs}
\alias{extract_chip_IDs}
\title{Extract the chip IDs to be used during job execution}
\usage{
extract_chip_IDs(xml_dir)
}
\arguments{
\item{xml_dir}{Directory of the XML file}
}
\value{
UIDs of the chip(s)
}
\description{
Extract the chip IDs to be used during job execution
}
\seealso{
Other database related: 
\code{\link{connect_mongoDB}()},
\code{\link{extract_files_to_process}()},
\code{\link{extract_gate_names}()},
\code{\link{extract_gates_to_process}()},
\code{\link{extract_jobtype}()},
\code{\link{find_all_attributes_in_EDL}()},
\code{\link{find_server_path}()},
\code{\link{find_valid_group_chip_IDs}()},
\code{\link{query_mongoDB}()}
}
\concept{database related}
\keyword{internal}
