% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_DBrelated.R
\name{extract_chipIDs_from_groupEDL}
\alias{extract_chipIDs_from_groupEDL}
\title{Title}
\usage{
extract_chipIDs_from_groupEDL(EDL)
}
\arguments{
\item{EDL}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
