% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_MetadataScanPosition.R
\name{extract_cycle_sequence_from_MethodHistory}
\alias{extract_cycle_sequence_from_MethodHistory}
\title{extract_cycle_sequence_from_MethodHistory}
\usage{
extract_cycle_sequence_from_MethodHistory(MethodHistory)
}
\arguments{
\item{MethodHistory}{}
}
\value{

}
\description{
extract_cycle_sequence_from_MethodHistory
}
