% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImportBlobFiles.R
\name{extract_enabled_positions}
\alias{extract_enabled_positions}
\title{extract_enabled_positions}
\usage{
extract_enabled_positions(single_pos_entity)
}
\arguments{
\item{single_pos_entity}{}
}
\value{

}
\description{
extract_enabled_positions
}
