% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_Cutoff.R, R/helpers_DBrelated.R
\name{extract_gate_metadata}
\alias{extract_gate_metadata}
\title{extracts gates metadata}
\usage{
extract_gate_metadata(chip_IDs)

extract_gate_metadata(chip_IDs)
}
\arguments{
\item{chip_IDs}{}

\item{gate_limslager}{}
}
\value{



}
\description{
This functions takes the return object of a mongo-collection limslager query for gate_IDs. The MetaData including gate_ID and the corresponding gateName, segment_ID, ParentGate_ID, x_marker_UID and y_marker_UID were combined and returned as dataframe.
}
\keyword{internal}
