% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_Cutoff.R
\name{extract_gate_name}
\alias{extract_gate_name}
\title{extracts name of gates}
\usage{
extract_gate_name(gate_limslager)
}
\arguments{
\item{gate_limslager}{}
}
\value{

}
\description{
extracts name of gates
}
\keyword{internal}
