% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{extract_gate_names}
\alias{extract_gate_names}
\title{Extract the names of the gates used for the job}
\usage{
extract_gate_names(xml_dir)
}
\arguments{
\item{xml_dir}{Directory of the XML file}
}
\value{
Gate names
}
\description{
Extract the names of the gates used for the job
}
\seealso{
Other database related: 
\code{\link{connect_mongoDB}()},
\code{\link{extract_chip_IDs}()},
\code{\link{extract_files_to_process}()},
\code{\link{extract_gates_to_process}()},
\code{\link{extract_jobtype}()},
\code{\link{find_server_path}()},
\code{\link{query_mongoDB}()}
}
\concept{database related}
\keyword{internal}
