% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImportBlobFiles.R
\name{extract_image_path_from_blob_parameter}
\alias{extract_image_path_from_blob_parameter}
\title{extract_image_path_from_blob_parameter}
\usage{
extract_image_path_from_blob_parameter(blob_parameter)
}
\arguments{
\item{blob_parameter}{}
}
\description{
extract_image_path_from_blob_parameter
}
\keyword{internal}
