% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImageProcessing.R
\name{extract_image_resolution}
\alias{extract_image_resolution}
\title{reads the resolution attribute of an image object}
\usage{
extract_image_resolution(image_data)
}
\arguments{
\item{filepath}{}
}
\value{

}
\description{
reads the resolution attribute of an image object
}
