% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{extract_information_from_RService_XML}
\alias{extract_information_from_RService_XML}
\title{Get relevant information from the input XML}
\usage{
extract_information_from_RService_XML(xml_dir)
}
\arguments{
\item{xml_dir}{Path to the parsed XML}
}
\value{
Relevant information: Job UID, job type, all file and gate names
}
\description{
Get relevant information from the input XML
}
\keyword{internal}
