% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_Cutoff.R
\name{extract_Marker_Cutoff}
\alias{extract_Marker_Cutoff}
\title{extracts VizMarkers Gate High and Low Values}
\usage{
extract_Marker_Cutoff(VizMarker_limslager)
}
\arguments{
\item{VizMarker_limslager}{}
}
\value{

}
\description{
extracts VizMarkers Gate High and Low Values
}
\keyword{internal}
