% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_Cutoff.R
\name{extract_ObjRef_chip_limslager}
\alias{extract_ObjRef_chip_limslager}
\title{extracts ObjRefs Type and UIDs}
\usage{
extract_ObjRef_chip_limslager(chip_limslager)
}
\arguments{
\item{chip_limslager}{}
}
\value{

}
\description{
extracts ObjRefs Type and UIDs
}
\keyword{internal}
