% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImportBlobFiles.R
\name{extract_parameter_from_BLOB}
\alias{extract_parameter_from_BLOB}
\title{extract_parameter_from_BLOB}
\usage{
extract_parameter_from_BLOB(image_path, blob_filename)
}
\arguments{
\item{blob_filename}{}
}
\value{

}
\description{
extract_parameter_from_BLOB
}
