% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{extract_paths_of_fl_files}
\alias{extract_paths_of_fl_files}
\title{Extract the paths of FL value files used for this job}
\usage{
extract_paths_of_fl_files(xml_dir)
}
\arguments{
\item{xml_dir}{Directory of the XML file}
}
\value{
File Paths of the FL value files
}
\description{
Extract the paths of FL value files used for this job
}
\keyword{internal}
