% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{extract_paths_of_gates}
\alias{extract_paths_of_gates}
\title{Extract the paths of the gate files used for the job}
\usage{
extract_paths_of_gates(xml_dir)
}
\arguments{
\item{xml_dir}{Directory of the XML file}
}
\value{
File paths for the gate files
}
\description{
Extract the paths of the gate files used for the job
}
