% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_Cutoff.R
\name{extract_VizMarker_ID}
\alias{extract_VizMarker_ID}
\title{extracts CytometryDatavizParametersMarker UID}
\usage{
extract_VizMarker_ID(DataViz_limslager)
}
\arguments{
\item{DataViz_ID}{}
}
\value{

}
\description{
extracts CytometryDatavizParametersMarker UID
}
\keyword{internal}
