% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_Cutoff.R
\name{extract_xml_value}
\alias{extract_xml_value}
\title{returns values of all nodes at a given xpath command}
\usage{
extract_xml_value(xml_node, find_string)
}
\arguments{
\item{find_string}{}
}
\value{

}
\description{
returns values of all nodes at a given xpath command
}
\examples{
find_string <- "//*/SpecificParameter[@Name='Parentsegmentation']"
find_string <- "//*/SpecificParameter[@Name='Parentgate']"
find_string <- "//*/SpecificParameter[@Name='x_marker_UID']"
find_string <- "//*/SpecificParameter[@Name='y_marker_UID']"


}
\keyword{internal}
