% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_Cutoff.R
\name{filter_ObjRef_Gate}
\alias{filter_ObjRef_Gate}
\title{filters Gate UIDs in ObjRef}
\usage{
filter_ObjRef_Gate(ObjRef)
}
\arguments{
\item{ObjRef}{}
}
\value{

}
\description{
filters Gate UIDs in ObjRef
}
\keyword{internal}
