% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_DBrelated.R
\name{find_chip_path}
\alias{find_chip_path}
\title{searches for a chip_ID}
\usage{
find_chip_path(ChannelID = "M583054", server_path = NULL)
}
\arguments{
\item{ChannelID}{chracter if the chip_ID}

\item{server_path}{vector containing characters of server paths}
}
\description{
finds the folder path to a related chip_ID in a list of server paths
}
\examples{
\donttest{

server <- find_server_path()
chip_path <- find_chip_path("'M1153570", server$server_path)
}
}
\keyword{internal}
