% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_DBrelated.R
\name{find_RService_XML_on_imageserver}
\alias{find_RService_XML_on_imageserver}
\title{finds the file path to RJob-xml}
\usage{
find_RService_XML_on_imageserver(chipID, segmentID, RJobID)
}
\arguments{
\item{chipID}{a character of chipID}

\item{segmentID}{a charceter of segmentID}

\item{RJobID}{a character of RJobID}
}
\value{
lists list of server_path's, chip_path, XML_path
}
\description{
querry mongoDB for the path to a given chipID in the list of available serverpaths and creates a final string to the folder where RJob XML can be found
}
\examples{
\donttest{
#need access to mongoDB
# chip in Leipzig
chipID <- "M186808"
segementID = "EZKL110986"
RJobID <- "EZKL113783"

find_RService_XML_on_imageserver(chipID,segementID,RJobID)
}
}
\seealso{
Other RJob mainscript: 
\code{\link{copy_file_positionCSV}()},
\code{\link{create_XML_filename}()},
\code{\link{load_mainscript_output}()}
}
\concept{RJob mainscript}
\keyword{internal}
