% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImportBlobFiles.R
\name{find_scan_basepath}
\alias{find_scan_basepath}
\title{find_scan_basepath}
\usage{
find_scan_basepath(scan_IDs)
}
\arguments{
\item{scan_IDs}{}
}
\value{

}
\description{
find_scan_basepath
}
