% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_DBrelated.R
\name{find_server_path}
\alias{find_server_path}
\title{finds available server storing images generated by ZKWapp}
\usage{
find_server_path()
}
\value{
a df containing flag, EDLName and path to the server
}
\description{
finds available server storing images generated by ZKWapp
}
\examples{
\donttest{
find_server_path()
}
}
\seealso{
Other database related: 
\code{\link{check_if_chip_data_exist}()},
\code{\link{connect_mongoDB}()},
\code{\link{extract_chipIDs_from_groupEDL}()},
\code{\link{find_chip_path}()},
\code{\link{find_valid_group_chip_IDs}()},
\code{\link{get_EDL_from_query_result}()}

Other database related: 
\code{\link{check_if_chip_data_exist}()},
\code{\link{connect_mongoDB}()},
\code{\link{extract_chipIDs_from_groupEDL}()},
\code{\link{find_chip_path}()},
\code{\link{find_valid_group_chip_IDs}()},
\code{\link{get_EDL_from_query_result}()}
}
\concept{database related}
\keyword{internal}
