% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_DBrelated.R
\name{get_df_from_query_result}
\alias{get_df_from_query_result}
\title{get_df_from_query_result}
\usage{
get_df_from_query_result(query_result)
}
\arguments{
\item{query_result}{}
}
\value{

}
\description{
get_df_from_query_result
}
\keyword{internal}
