% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_DBrelated.R
\name{get_EDLs_of_queryResult}
\alias{get_EDLs_of_queryResult}
\title{extracts all EDL strings from a mongoDB query result}
\usage{
get_EDLs_of_queryResult(query_result)
}
\arguments{
\item{query_result}{list containing the result of a mongoDB-query}

\item{IDs}{chracter if IDs}
}
\value{
a character vector containing EDL strings
}
\description{
internally checks if EDL exist and can be read and if not returns a error-message.string
}
\examples{
query_result <- list(
result = list(structure(list(
EDL = "<?xml version=\"1.0\"?>\r\n<!-- ",
EDLType = "ChipGroup",
query = "UID_P1761451_limslager"), class = "data.frame", row.names = 1L)),
error_message = character(0))

EDL <- query_result\%>\%
get_EDLs_of_queryResult()

query_result <- list(
result = list(structure(list(
EDL = "EDL blabla",
EDLType = "ChipGroup",
query = "UID_P1761451_limslager"), class = "data.frame", row.names = 1L)),
error_message = character(0))

query_result\%>\%
get_EDLs_of_queryResult()

"test"\%>\%
get_EDLs_of_queryResult()
}
\seealso{
Other mongoDB: 
\code{\link{query_mongoDB}()}
}
\concept{mongoDB}
\keyword{internal}
