% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_DBrelated.R
\name{get_enabled_positions}
\alias{get_enabled_positions}
\title{get_enabled_positions}
\usage{
get_enabled_positions(chip_ID)
}
\arguments{
\item{chip_ID}{}
}
\description{
get_enabled_positions
}
\keyword{internal}
