% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImportBlobFiles.R
\name{get_enabled_positions_from_positions_list}
\alias{get_enabled_positions_from_positions_list}
\title{get_enabled_positions_from_positions_list}
\usage{
get_enabled_positions_from_positions_list(positions_list)
}
\arguments{
\item{positions_list}{}
}
\description{
get_enabled_positions_from_positions_list
}
\keyword{internal}
