% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_errorHandling.R
\name{get_errorDF_queryResults}
\alias{get_errorDF_queryResults}
\title{checks mongoDB query result for missing entities}
\usage{
get_errorDF_queryResults(query_result, error_string = "error_", print = FALSE)
}
\arguments{
\item{query_result}{list, generated by query_mongoDB, containing a list result and error_message}

\item{error_string}{character of a string defining an error}
}
\value{
dataframe of error_ID and error_text
}
\description{
returns a dataframe with error_IDs and error_text generated by the query_mongoDB().
}
\examples{
\donttest{
query_result <- c("P1761451","re","t")\%>\%query_UID_limslager()
error_df <- query_result\%>\%get_errorDF_queryResults()
error_df <- query_result\%>\%get_errorDF_queryResults(print = TRUE)

"P1761451"\%>\%query_UID_limslager()\%>\%get_errorDF_queryResults()
}
}
\seealso{
Other errorHandling: 
\code{\link{check_EDLs_strings}()},
\code{\link{check_mappingResultLists}()},
\code{\link{detect_error_inCharacterVector}()},
\code{\link{is_error_in_EDLs}()},
\code{\link{is_error_in_mappingResult}()},
\code{\link{print_queryResult_error}()},
\code{\link{split_errorText_toDF}()}
}
\concept{errorHandling}
\keyword{internal}
