% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{get_gate_ID_of_AllGate}
\alias{get_gate_ID_of_AllGate}
\title{Title}
\usage{
get_gate_ID_of_AllGate(chip_ID, segment_ID)
}
\arguments{
\item{segment_ID}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
