% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{get_identityID_from_EDL}
\alias{get_identityID_from_EDL}
\title{Title}
\usage{
get_identityID_from_EDL(EDL)
}
\arguments{
\item{EDL}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
