% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_DevelTools.R
\name{get_package_dependencies}
\alias{get_package_dependencies}
\title{create df of package dependencies}
\usage{
get_package_dependencies(
  exact_version = "mongolite",
  ignore_folder = "devel",
  ignore_package = "RJobTissueArea"
)
}
\arguments{
\item{exact_version}{character vector of package names where exact version is obligatory}

\item{ignore_folder}{character vector of folder names in pkg directory which are build-ignored}

\item{ignore_package}{character vector of package names to remove from final list}
}
\value{

}
\description{
create df of package dependencies
}
\examples{
get_package_dependencies(
exact_version = c("mongolite","locfit","EBImage"),
ignore_folder <- c("devel"),
ignore_package <- "RJobTissueArea"
 )
}
