% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_MetadataScanPosition.R
\name{get_pos_colnames_in_MethodHistory}
\alias{get_pos_colnames_in_MethodHistory}
\title{get_pos_colnames_in_MethodHistory}
\usage{
get_pos_colnames_in_MethodHistory(MH)
}
\arguments{
\item{MH}{}
}
\value{

}
\description{
get_pos_colnames_in_MethodHistory
}
