% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{get_sampleType_from_MethodHistory}
\alias{get_sampleType_from_MethodHistory}
\title{Title}
\usage{
get_sampleType_from_MethodHistory(MethodHistory)
}
\arguments{
\item{MethodHistory}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
