% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_MetadataScanPosition.R
\name{get_sampleType_of_chipIDs}
\alias{get_sampleType_of_chipIDs}
\title{returns the sampleType of chip_IDs}
\usage{
get_sampleType_of_chipIDs(chip_IDs)
}
\arguments{
\item{chip_IDs}{character vector of chip_IDs}
}
\value{
list of sampleTypes named by chipIDs
}
\description{
returns the sampleType of chip_IDs
}
\examples{
\donttest{
data("chip_IDs")
sampleType <- get_sampleType_of_chip_IDs(chip_IDs)
test <- get_sampleType_of_chipIDs("test")
}
}
\concept{MetadataScanPosition}
\keyword{internal}
