% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{get_segment_ID_of_chipID}
\alias{get_segment_ID_of_chipID}
\title{Title}
\usage{
get_segment_ID_of_chipID(chip_ID)
}
\arguments{
\item{chip_ID}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
