% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_errorHandling.R
\name{handle_trycache_error}
\alias{handle_trycache_error}
\title{standardizes try-chache error function}
\usage{
handle_trycache_error(err, task, enable.quit = FALSE, output_dir = getwd())
}
\arguments{
\item{err}{character specifying the error generated and assigned by the tryCatch function}

\item{task}{character specifying the task, processed within the tryCatch}

\item{enable.quit}{logical if quit without saving}
}
\value{
prints a message into console and appends the error.txt-file in the \code{output_dir}
}
\description{
creates a message string consisting of the error message produced in the try chache and a task defined by the user.
Writes the message into the console and into a textfile named by "error_message.txt", located in the \code{output_dir}.
if \code{enable.quit} is set \code{TRUE} R execution will quit without saving
}
\examples{

output.dir <- getwd()
enable.quit<- FALSE
task<-"test"

testthat::expect_warning(
tryCatch({
log(-1)
 }, error = function(err) {handle_trycache_error(err,task,enable.quit)})
)
testthat::expect_warning(
tryCatch({
task<-"test_2"
log(-1)
 }, error = function(err) {handle_trycache_error(err,task,enable.quit)})
)
testthat::expect_warning(
 tryCatch({
log(-1)
 }, error = function(err) {handle_trycache_error(err,task,enable.quit)})
)
}
\seealso{
Other develTools: 
\code{\link{get_nonError_index}()}
}
\concept{develTools}
\keyword{internal}
