% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{import_all_fl_files}
\alias{import_all_fl_files}
\title{Import multiple .csv files}
\usage{
import_all_fl_files(all_file_paths, csv_type = "FLCSV", chip_IDs, jobtype)
}
\arguments{
\item{all_file_paths}{Valid paths to all files that are to be imported}

\item{csv_type}{Type of the files}

\item{chip_IDs}{UIDs of the chips to which the listed files belong}

\item{jobtype}{Type of the job that is to be executed}
}
\value{
List with all \code{.csv} files imported as \code{data.table}
}
\description{
Import multiple .csv files
}
\keyword{internal}
