% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{import_all_gate_files}
\alias{import_all_gate_files}
\title{Import a list of gate files}
\usage{
import_all_gate_files(all_file_paths, gatenames, chip_ID)
}
\arguments{
\item{all_file_paths}{Valid file paths}

\item{gatenames}{Names of the gates}

\item{chip_ID}{Chip UID}
}
\value{
List of gate files as \code{data.table}s
}
\description{
Import a list of gate files
}
