% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{import_customer_form}
\alias{import_customer_form}
\title{Title}
\usage{
import_customer_form(chipCSV, ColumnsToSelect, header_row = 3)
}
\arguments{
\item{header_row}{}
}
\value{

}
\description{
Title
}
