% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{import_R_Service_XML_file}
\alias{import_R_Service_XML_file}
\title{Import XML file containing job description and file paths}
\usage{
import_R_Service_XML_file(
  XML_filepath,
  output_dir,
  RJob_ID,
  chip_ID,
  segment_ID
)
}
\arguments{
\item{XML_filepath}{Path to the XML file}

\item{output_dir}{Output directory for the parsed XML file}

\item{RJob_ID}{R Job UID}

\item{chip_ID}{Chip UID}

\item{segment_ID}{Segmentation job UID}
}
\value{
File path of the parsed XML
}
\description{
Import XML file containing job description and file paths
}
\keyword{internal}
