% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{import_single_fl_file}
\alias{import_single_fl_file}
\title{Import one .csv file}
\usage{
import_single_fl_file(file_path, csv_type, chip_ID, jobtype)
}
\arguments{
\item{file_path}{Valid path to the file}

\item{csv_type}{Type of the file, e.g. "FLCSV"}

\item{chip_ID}{UID of the chip to which this file belongs}

\item{jobtype}{Type of the job that is being executed}
}
\value{
A tidy \code{.csv} file imported as \code{data.table}
}
\description{
Import one .csv file
}
\keyword{internal}
