% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{import_single_gate_file}
\alias{import_single_gate_file}
\title{Import a single gate file}
\usage{
import_single_gate_file(file_path)
}
\arguments{
\item{file_path}{A valid file path}
}
\value{
Gate file as \code{data.table}
}
\description{
Import a single gate file
}
