% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{import_ZKW_rawdata_to_flvalues}
\alias{import_ZKW_rawdata_to_flvalues}
\title{Title}
\usage{
import_ZKW_rawdata_to_flvalues(
  chip_path,
  chip_ID,
  segment_ID,
  gate_ID,
  markernames = "CD4"
)
}
\arguments{
\item{markernames}{}
}
\value{

}
\description{
Title
}
