% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_errorHandling.R
\name{is_error_in_mappingResult}
\alias{is_error_in_mappingResult}
\title{checks for error messages in result list}
\usage{
is_error_in_mappingResult(resultList)
}
\arguments{
\item{resultList}{a result list containing error messages, potentially}
}
\value{
logical if list element is a error
}
\description{
checks for error messages in result list
}
\examples{
resultList <- list(df1 = data.frame(col1=c(1,2,3),col2=c(1,2,3)),
df2 = data.frame(col1=c(1,2,3),col2=c(1,2,3)),
error="error_text")

resultList\%>\%is_error_in_mappingResult()
}
\seealso{
Other errorHandling: 
\code{\link{check_EDLs_strings}()},
\code{\link{check_mappingResultLists}()},
\code{\link{detect_error_inCharacterVector}()},
\code{\link{get_errorDF_queryResults}()},
\code{\link{is_error_in_EDLs}()},
\code{\link{print_queryResult_error}()},
\code{\link{split_errorText_toDF}()}
}
\concept{errorHandling}
\keyword{internal}
