% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImportBlobFiles.R
\name{list_all_image_files}
\alias{list_all_image_files}
\title{list_all_image_files}
\usage{
list_all_image_files(chip_IDs, type)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
list_all_image_files
}
