% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImportBlobFiles.R
\name{list_BlobFileName_in_filepath}
\alias{list_BlobFileName_in_filepath}
\title{list_BlobFileName_in_filepath}
\usage{
list_BlobFileName_in_filepath(image_filepath)
}
\arguments{
\item{image_filepath}{}
}
\value{

}
\description{
list_BlobFileName_in_filepath
}
