% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImportBlobFiles.R
\name{list_posFolders_in_ScanBasePath}
\alias{list_posFolders_in_ScanBasePath}
\title{list_posFolders_in_ScanBasePath}
\usage{
list_posFolders_in_ScanBasePath(ScanBasePath)
}
\arguments{
\item{ScanBasePath}{}
}
\value{

}
\description{
list_posFolders_in_ScanBasePath
}
\keyword{internal}
