% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{load_mainscript_output}
\alias{load_mainscript_output}
\title{loads all objects generated by the mainscript}
\usage{
load_mainscript_output(chip_ID, RJob_ID, input.dir = NULL)
}
\arguments{
\item{chip_ID}{character}

\item{RJob_ID}{character}

\item{input.dir}{character defining directory of the data to load, if set to
null standard is subdirectory data/chip_ID/mainscript}
}
\description{
loads objects present in the Renvironment of ZKWapp after mainscript
exceution, run after execute_RJob_mainscript() to load all the objects
}
\details{
input, flvalues, gates, xml.dir, , UID, jobtype, gatenames, chip.IDs
and in addition PositionCSV and chip_path are loaded into the
user-R-environment
}
\examples{
\donttest{
chip_ID <- "M1153570"
segment_ID <- "E1549294"
RJob_ID <- "E1568014"

# execute_RJob_mainscript(chip_ID,segment_ID,RJob_ID,NULL)
load_mainscript_output(chip_ID, RJob_ID, NULL)
}
}
\seealso{
Other RJob mainscript: 
\code{\link{copy_file_positionCSV}()},
\code{\link{create_XML_filename}()},
\code{\link{find_RService_XML_on_imageserver}()}
}
\concept{RJob mainscript}
\keyword{MainScript}
