% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{merge_all_input_files_to_one_list_and_export}
\alias{merge_all_input_files_to_one_list_and_export}
\title{Combine all input files and meta information to one list and export it}
\usage{
merge_all_input_files_to_one_list_and_export(
  gate_files,
  fl_files,
  XML_file_content,
  output_dir,
  chip_ID,
  RJob_ID
)
}
\arguments{
\item{gate_files}{All gate files}

\item{fl_files}{All FL value files}

\item{XML_file_content}{Content of the R Service XML file containing the Chip and R Job UIDs}

\item{output_dir}{Output directory}

\item{chip_ID}{Chip UID}

\item{RJob_ID}{R Job UId}
}
\description{
Combine all input files and meta information to one list and export it
}
\keyword{internal}
