% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{merge_fl_and_gate_files}
\alias{merge_fl_and_gate_files}
\title{Merge chip fl value file and corresponding gate files}
\usage{
merge_fl_and_gate_files(gate_files, curr_fl_file, chip_ID)
}
\arguments{
\item{gate_files}{All gate files from the current job}

\item{curr_fl_file}{The fl value file of a chip}

\item{chip_ID}{The chip UID}
}
\value{
A list containing one element for each gate that corresponds to this
chip UID. Each list element is itself a list element containing the FL
values of all the chips within this gate and further information as
produced by \code{\link{import_all_fl_files}}.
}
\description{
Merge chip fl value file and corresponding gate files
}
