% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_errorHandling.R
\name{print_queryResult_error}
\alias{print_queryResult_error}
\title{prints an overview of entities where mongoDB query serves no return}
\usage{
print_queryResult_error(error_df)
}
\arguments{
\item{error_df}{data.frame or NULL, output generated by get_errorDF_queryResults()}
}
\value{
print to console
}
\description{
prints an overview of entities where mongoDB query serves no return
}
\examples{
\donttest{
query_result <- c("P1761451","re","t")\%>\%query_UID_limslager()
query_result\%>\%get_errorDF_queryResults()\%>\%print_queryResult_error
"P1761451"\%>\%query_UID_limslager()\%>\%
get_errorDF_queryResults()\%>\%print_queryResult_error()
}
}
\seealso{
Other errorHandling: 
\code{\link{check_EDLs_strings}()},
\code{\link{check_mappingResultLists}()},
\code{\link{detect_error_inCharacterVector}()},
\code{\link{get_errorDF_queryResults}()},
\code{\link{is_error_in_EDLs}()},
\code{\link{is_error_in_mappingResult}()},
\code{\link{split_errorText_toDF}()}
}
\concept{errorHandling}
\keyword{internal}
