% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImageProcessing.R
\name{process_tissue_detection_workflow}
\alias{process_tissue_detection_workflow}
\title{process_tissue_detection_workflow}
\usage{
process_tissue_detection_workflow(m.data, cellres, sigma, threshold, window)
}
\arguments{
\item{window}{}
}
\description{
process_tissue_detection_workflow
}
