% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImageProcessing.R
\name{process_TissueDetection}
\alias{process_TissueDetection}
\title{process_TissueDetection}
\usage{
process_TissueDetection(
  image_groups,
  output_dir,
  sigma = 15,
  threshold = 2,
  window = 10,
  attenuation = 0.01,
  noiseReduction = TRUE,
  plot_image = TRUE,
  result_ID = ""
)
}
\arguments{
\item{result_ID}{}
}
\description{
process_TissueDetection
}
\examples{
\donttest{
output_dir <- "devel/data/data_output"
result_ID <- "P1761451"
group_ID <- "P1761451"
RJobTissueArea:::create_working_directory(output_dir)
chip_IDs <- find_valid_group_chip_IDs(group_ID)
ScanHistory <- create_ScanHistory_extended(chip_IDs,
output_dir,
result_ID = group_ID)
image_groups <- create_hdr_image_groups(ScanHistory)
sigma = 15
threshold = 2
window = 10
attenuation = 0.01
noiseReduction = FALSE
plot_image = TRUE
}
}
