% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImportBlobFiles.R
\name{query_filterset_of_scanIDs}
\alias{query_filterset_of_scanIDs}
\title{query_filterset_of_scanIDs}
\usage{
query_filterset_of_scanIDs(scan_IDs)
}
\arguments{
\item{scan_IDs}{}
}
\value{

}
\description{
query_filterset_of_scanIDs
}
