% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_DBrelated.R
\name{query_mongoDB}
\alias{query_mongoDB}
\title{query_mongoDB()}
\usage{
query_mongoDB(mongo_collection, attribute_name, attribute_value)
}
\arguments{
\item{mongo_collection}{character, name of the ZKW mongoDB collection, "limslager" or "limsproc"}

\item{attribute_name}{character, name of the EDL attribute}

\item{attribute_value}{character, value of the EDL attribute}
}
\value{
list of df: result and vector: error_message
}
\description{
This Function performs for you a query in a ZKW-mongoDB-collection and tries to find those EDL objects for which a certain attribute matches a value.
}
\details{
You need to specify
\enumerate{
\item which collection you want to look at: "limslager" or "limsprroc" and
\item what you want to look at, meaning the name of an attribute and
\item what you want to find, the value of the attribute.
}

The function calls connect_mongoDB() to provide the external pointer to the DB-collection.
It creates a json string of the query-command, performs the query for all json strings and collects the query results.
A string containing an "mongo_collection", "attribute_name" and "attribute_value" is added to each results.
It further records querries which yielded no result, meaning that there is no entry in the database and
creates for those a vercor error messages containing all error messages of the query.
All results where combined to a list of the result-list and the error message is returned.
}
\examples{
\donttest{

query_result <- query_mongoDB(
  mongo_collection = "limsproc",
  attribute_name = "UID",
  attribute_value = "M856222"
)

query_result <- query_mongoDB(
  mongo_collection = "limslager",
  attribute_name = "EDLType",
  attribute_value = c("Antibody Stock Solution", "Antibody Stock Solution")
)
}
}
\seealso{
Other database related: 
\code{\link{check_if_chip_data_exist}()},
\code{\link{connect_mongoDB}()},
\code{\link{extract_chipIDs_from_groupEDL}()},
\code{\link{extract_chip_IDs}()},
\code{\link{extract_files_to_process}()},
\code{\link{extract_gate_names}()},
\code{\link{extract_gates_to_process}()},
\code{\link{extract_jobtype}()},
\code{\link{find_all_attributes_in_EDL}()},
\code{\link{find_chip_path}()},
\code{\link{find_server_path}()},
\code{\link{find_valid_group_chip_IDs}()},
\code{\link{get_EDL_from_query_result}()}
}
\concept{database related}
\keyword{internal}
