% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_DBrelated.R
\name{query_segment_ID}
\alias{query_segment_ID}
\title{Title}
\usage{
query_segment_ID(segment_IDs)
}
\arguments{
\item{segment_IDs}{}
}
\value{
data.frame
}
\description{
Title
}
\examples{
segment_IDs <- c("EZKL122811","EZKL207112","EZKL289562")
}
\keyword{internal}
