% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{query_UID_limslager}
\alias{query_UID_limslager}
\title{Title}
\usage{
query_UID_limslager(chip_IDs)
}
\arguments{
\item{chip_path}{}
}
\value{
result
}
\description{
Title
}
\examples{
chip_path <- c("\\\\\\\\intern.chipcytometry.com\\\\imagedata\\\\leipzig_volume0\\\\M986054")
collect_segments_metadata(chip_path)
}
