% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_DBrelated.R
\name{query_UID_limsproc}
\alias{query_UID_limsproc}
\title{query_UID_limsproc}
\usage{
query_UID_limsproc(chip_IDs)
}
\arguments{
\item{chip_IDs}{}
}
\description{
query_UID_limsproc
}
\keyword{internal}
