% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImportBlobFiles.R
\name{read_binary_image_as_matrix}
\alias{read_binary_image_as_matrix}
\title{read_binary_image_as_matrix}
\usage{
read_binary_image_as_matrix(image_path, blob_filename)
}
\arguments{
\item{blob_filename}{}
}
\description{
read_binary_image_as_matrix
}
\keyword{internal}
