% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImportBlobFiles.R
\name{read_binary_image_as_vector}
\alias{read_binary_image_as_vector}
\title{read_binary_image_as_vector}
\usage{
read_binary_image_as_vector(image_path, blob_filename)
}
\arguments{
\item{blob_filename}{}
}
\value{

}
\description{
read_binary_image_as_vector
}
\keyword{internal}
