% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImageProcessing.R
\name{read_data_sum_as_matrix}
\alias{read_data_sum_as_matrix}
\title{read_data_sum_as_matrix}
\usage{
read_data_sum_as_matrix(filepath)
}
\arguments{
\item{filepath}{}
}
\value{

}
\description{
read_data_sum_as_matrix
}
