% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImageProcessing.R
\name{read_data_sum_image_resolution}
\alias{read_data_sum_image_resolution}
\title{read_data_sum_image_resolution}
\usage{
read_data_sum_image_resolution(filepath)
}
\arguments{
\item{filepath}{}
}
\value{

}
\description{
read_data_sum_image_resolution
}
