% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImportBlobFiles.R
\name{read_image_binary_file}
\alias{read_image_binary_file}
\title{read_image_binary_file}
\usage{
read_image_binary_file(blob_parameter)
}
\arguments{
\item{blob_parameter}{}
}
\value{

}
\description{
read_image_binary_file
}
