% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_DevelTools.R
\name{read_result_file}
\alias{read_result_file}
\title{load a result file and check columns}
\usage{
read_result_file(result_df, result_filename)
}
\arguments{
\item{result_filename}{}
}
\value{

}
\description{
load a result file and check columns
}
\examples{
result_df <- tibble::tibble(
group_ID = character(0), # group_ID
chip_ID = character(0), #chip_ID
pos_ID = numeric(0), #pos_ID
sigma = numeric(0),
threshold = numeric(0),
GS_window = numeric(0),
perc_TissueArea = double(0),
TissueArea_mm2 = double(0)
)
}
\keyword{internal}
