% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImportBlobFiles.R
\name{read_ScanHistory}
\alias{read_ScanHistory}
\title{read_ScanHistory}
\usage{
read_ScanHistory(group_ID, output_dir)
}
\arguments{
\item{output_dir}{}
}
\value{

}
\description{
read_ScanHistory
}
\keyword{internal}
