% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImportBlobFiles.R
\name{read_XML_BLOB_parameter}
\alias{read_XML_BLOB_parameter}
\title{read_XML_BLOB_parameter}
\usage{
read_XML_BLOB_parameter(image_path, blob_filename)
}
\arguments{
\item{blob_filetype}{}
}
\description{
read_XML_BLOB_parameter
}
\keyword{internal}
