% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImageProcessing.R
\name{reduce_noise}
\alias{reduce_noise}
\title{determines noise and sets FLvalues to zero}
\usage{
reduce_noise(data)
}
\arguments{
\item{data}{vector or metrix of values}
}
\value{
vector of metric of values with noise values set to zero
}
\description{
\itemize{
\item calculates density of FLvalues and determines all peaks
\item defines the first density peak as noise
\item noise threshold equals the last FLvalue of the first peak in the density curve
\item sets all FLvalues below noise threshold to zero
}
}
\examples{
\donttest{
group_ID <- "P1761451"
output_dir <- "data_output"
RJobTissueArea:::create_working_directory(output_dir)
chip_IDs <- find_valid_group_chip_IDs(group_ID)
ScanHistory <- create_ScanHistory_extended(chip_IDs,output_dir,group_ID)
image_groups <- create_hdr_image_groups(ScanHistory)
image_group_list<-image_groups$data[[51]]
data_sum <- calculate_data_sum(image_group_list)
data_attenuate <- attenuate_FLpeakValues(data_sum,0.01)
data_noi <- reduce_noise(data_attenuate)
m.data <- convert_image_vector_to_matrix(data_noi)
cellres <- extract_image_resolution(data_sum)
grey_values <- create_pixmap_greyvalues(m.data,cellres)
grey_values\%>\%
imager::as.cimg()\%>\%
plot(main = "original dataSum")
}
}
\keyword{internal}
