% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_MetadataScanPosition.R
\name{return_segment_metadata}
\alias{return_segment_metadata}
\title{Title}
\usage{
return_segment_metadata(chip_IDs)
}
\arguments{
\item{chip_IDs}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
