% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RJobTissueArea-package.R
\docType{package}
\name{RJobTissueArea-package}
\alias{RJobTissueArea}
\alias{RJobTissueArea-package}
\title{RJobTissueArea: RJobTissueArea}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Determination of chip area covered by tissue in selected
    positions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/juleZWK/RJobTissueArea}
  \item Report bugs at \url{https://github.com/juleZWK/RJobTissueArea/issues}
}

}
\author{
\strong{Maintainer}: Julia Ortmann \email{julia.ortmann@bruker.com}

}
\keyword{internal}
