% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_Cutoff.R
\name{search_chip_limslager}
\alias{search_chip_limslager}
\title{searches Chip_IDs in limslager}
\usage{
search_chip_limslager(chip_IDs)
}
\arguments{
\item{chip_IDs}{}
}
\value{

}
\description{
searches Chip_IDs in limslager
}
\examples{
chip_IDs <- c("M911981", "M911971")
}
\keyword{internal}
