% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_Cutoff.R
\name{search_CytometryDatavizParametersMarker_limslager}
\alias{search_CytometryDatavizParametersMarker_limslager}
\title{searches CytometryDatavizParametersMarker UID in limslager}
\usage{
search_CytometryDatavizParametersMarker_limslager(VizMarker_ID)
}
\arguments{
\item{Viz_Marker_ID}{}
}
\value{

}
\description{
searches CytometryDatavizParametersMarker UID in limslager
}
\examples{
VizMarker_ID <- c("D972732", "D972733", "D972734", "D972735", "D972736", "D972737", "D972787", "D972794",
"D972795", "D972796", "D972842", "D972844", "D972638", "D972639", "D972640", "D972683", "D972684",
"D972685", "D972790", "D972791", "D972793", "D972820", "D972843")
}
\keyword{internal}
