% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_Cutoff.R
\name{search_CytometryDatavizParametersSample_limslager}
\alias{search_CytometryDatavizParametersSample_limslager}
\title{searches CytometryDatavizParametersSample UID in limslager}
\usage{
search_CytometryDatavizParametersSample_limslager(DataViz_ID)
}
\arguments{
\item{DataViz_ID}{}
}
\value{

}
\description{
searches CytometryDatavizParametersSample UID in limslager
}
\examples{
DataViz_ID <- c("D972192","D972187")
}
\keyword{internal}
