% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_Cutoff.R
\name{search_gate_limslager}
\alias{search_gate_limslager}
\title{searches gate_IDs in limslager}
\usage{
search_gate_limslager(gate_IDs = Gates$gate_ID)
}
\arguments{
\item{gate_IDs}{}
}
\value{

}
\description{
searches gate_IDs in limslager
}
\keyword{internal}
