% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImportBlobFiles.R
\name{select_hdr_files}
\alias{select_hdr_files}
\title{select_hdr_files}
\usage{
select_hdr_files(result_files)
}
\arguments{
\item{results_files}{}
}
\value{

}
\description{
select_hdr_files
}
\keyword{internal}
