% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_MetadataScanPosition.R
\name{select_segment_ID}
\alias{select_segment_ID}
\title{Title}
\usage{
select_segment_ID(segment_IDs)
}
\arguments{
\item{segment_IDs}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
