% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImportBlobFiles.R
\name{select_valid_image_files}
\alias{select_valid_image_files}
\title{select_valid_image_files}
\usage{
select_valid_image_files(result_files, type = NULL)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
select_valid_image_files
}
