% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{set_attrs_full_input}
\alias{set_attrs_full_input}
\title{Save job metadata as attributes}
\usage{
set_attrs_full_input(full_input, tidy_XML_file_content, output_dir, RJob_ID)
}
\arguments{
\item{full_input}{List of all files to be saved}

\item{tidy_XML_file_content}{Metadata extracted from the R Service XML}

\item{output_dir}{Output directory}

\item{RJob_ID}{R Job UID}
}
\value{
\code{full_input} with attributes
}
\description{
Save job metadata as attributes
}
