% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_errorHandling.R
\name{split_errorText_toDF}
\alias{split_errorText_toDF}
\title{splits a error_message and returns a dataframe}
\usage{
split_errorText_toDF(error)
}
\arguments{
\item{error}{character vector of error messages}
}
\value{
dataframe containing error_ID and error_text
}
\description{
takes a character vector of error_messages and splits each character by "_".
extracts the second and third split product.
returns a dataframe containing error_ID and error_text
}
\examples{

error <- c("error_some text_M0815","Error_some further text_M0819")
split_errorText_toDF(error)

}
\seealso{
Other errorHandling: 
\code{\link{check_EDLs_strings}()},
\code{\link{check_mappingResultLists}()},
\code{\link{detect_error_inCharacterVector}()},
\code{\link{get_errorDF_queryResults}()},
\code{\link{is_error_in_EDLs}()},
\code{\link{is_error_in_mappingResult}()},
\code{\link{print_queryResult_error}()}
}
\concept{errorHandling}
\keyword{internal}
