% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImageProcessing.R
\name{summarize_result_df}
\alias{summarize_result_df}
\title{summarizes tissue area results per chip_ID}
\usage{
summarize_result_df(result_df)
}
\value{
a dataframe containing additional columns, \code{n_positions}, \code{perc_TissueArea, \code{TissueArea_mm2
}
\description{
summarizes tissue area results per chip_ID
}
\seealso{
Other ImageProcessing: 
\code{\link{calculate_data_sum}()}
}
\concept{ImageProcessing}
\keyword{internal}
