% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory.R
\name{target_factory}
\alias{target_factory}
\title{Tissue Area Analysis Factory}
\usage{
target_factory(group_ID, output_dir, sigma, threshold, window)
}
\arguments{
\item{window}{}
}
\value{
A list of target objects.
}
\description{
performs complete workflow of TissueAreaAnalysis
}
\details{
is targets-factory function which calls the functions supplied with this package and stores and tracks the output
Define 6 targets:
\enumerate{
\item Track the user-supplied input
\item loads valid chipIDs \code{load_valid_chipIDs()}
\item generate ScanHistory \code{generate_ScanHistory()}
\item creates input data \code{generate_InputImage()}
\item performs Image Processing \code{process_Images()}
\item calculates TissueArea \code{calculate_TissueArea()}
\item exports the result \code{export_TissueAreaResults()}
}
}
