% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{tidy_up_XML_file_content}
\alias{tidy_up_XML_file_content}
\title{Adapt the XML file content to the package needs}
\usage{
tidy_up_XML_file_content(XML_file_content)
}
\arguments{
\item{XML_file_content}{All relevant information from the RService XML as
produced by \code{\link{extract_information_from_RService_XML}}}
}
\value{
Tidied XML file content
}
\description{
For the gate names, "+" is replaced by "plus", "-" is replaced by "minus",
punctuation is replaced by underscore and non-unique names get a count.
Special characters need to be removed to produce valid file names in later
steps.
}
\keyword{internal}
