% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{write_lines_task}
\alias{write_lines_task}
\title{writes user lines of a task}
\usage{
write_lines_task(task, subtask = NULL, type = c("S", "E", "F"))
}
\arguments{
\item{task}{character of the task}

\item{subtask}{character of the subtask}

\item{type}{vcharacter speciefieing if to inform about the start or finish}
}
\value{
a console print produces by writeLines
}
\description{
writes the print message for Start or Completion of a task, subtask name supplied
}
\examples{
{
  write_lines_task("task", type = "S")
  write_lines_task("task", "subtask", "E")
}
}
\concept{programming helpers}
\keyword{internal}
