% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImageProcessing.R
\name{calculate_data_sum}
\alias{calculate_data_sum}
\title{calculate_data_sum}
\usage{
calculate_data_sum(image_list)
}
\arguments{
\item{image_group_list}{}
}
\description{
calculate_data_sum
}
\keyword{internal}
