% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImageProcessing.R
\name{calculate_TissueArea}
\alias{calculate_TissueArea}
\title{calculate_TissueArea}
\usage{
calculate_TissueArea(n_pixel, cellres)
}
\arguments{
\item{cellres}{}
}
\description{
calculate_TissueArea
}
\keyword{internal}
