% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_DBrelated.R
\name{check_if_chip_data_exist}
\alias{check_if_chip_data_exist}
\title{checks if chipID data folder exist}
\usage{
check_if_chip_data_exist(chip_IDs)
}
\arguments{
\item{chip_IDs}{character vector of chip_IDs}
}
\value{
character vector of chip_IDs for which data folder exist
}
\description{
Internally query for available Imageserver. Subsequent maps through all chip_IDs and check if a folder named by chip_IDs exist on the Server. Finally seperates the chip_IDs by folder existance and prints them seperatly. Returns only existing data chip_IDs.
}
\examples{
\donttest{
data(chip_IDs)
check_if_chip_data_exist(chip_IDs)
}

}
\seealso{
Other database related: 
\code{\link{connect_mongoDB}()},
\code{\link{extract_chipIDs_from_groupEDL}()},
\code{\link{extract_chip_IDs}()},
\code{\link{extract_files_to_process}()},
\code{\link{extract_gate_names}()},
\code{\link{extract_gates_to_process}()},
\code{\link{extract_jobtype}()},
\code{\link{find_all_attributes_in_EDL}()},
\code{\link{find_chip_path}()},
\code{\link{find_server_path}()},
\code{\link{find_valid_group_chip_IDs}()},
\code{\link{get_EDL_from_query_result}()},
\code{\link{query_mongoDB}()}
}
\concept{database related}
\keyword{internal}
