% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_DBrelated.R
\name{check_if_chip_data_exist}
\alias{check_if_chip_data_exist}
\title{Title}
\usage{
check_if_chip_data_exist(chip_IDs)
}
\arguments{
\item{chip_IDs}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
