% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_MetadataScanPosition.R
\name{create_MethodHistory_from_EDLs}
\alias{create_MethodHistory_from_EDLs}
\title{Title}
\usage{
create_MethodHistory_from_EDLs(EDLs)
}
\arguments{
\item{EDLs}{character vector containing EDL strings}
}
\value{

}
\description{
Title
}
\examples{
data("EDL_chipIDs_limslager")

MH <- create_MethodHistory_from_EDL(EDL_chipIDs_limslager)

MHtest <- create_MethodHistory_from_EDL(c("test",EDL_chipIDs_limslager))


}
\keyword{internal}
