% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImportBlobFiles.R
\name{create_MethodHistory_of_chipIDs}
\alias{create_MethodHistory_of_chipIDs}
\title{create_MethodHistory_of_chipIDs}
\usage{
create_MethodHistory_of_chipIDs(chip_IDs)
}
\arguments{
\item{chip_IDs}{}
}
\value{

}
\description{
create_MethodHistory_of_chipIDs
}
