% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImageProcessing.R
\name{create_plot_directory}
\alias{create_plot_directory}
\title{create_plot_directory}
\usage{
create_plot_directory(
  output_dir,
  sigma,
  threshold,
  window,
  attenuation,
  noiseReduction,
  chip_ID,
  pos_ID,
  suffix_resultfile
)
}
\arguments{
\item{pos_ID}{}
}
\description{
create_plot_directory
}
\keyword{internal}
