% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImportBlobFiles.R
\name{create_ScanHistory_extended}
\alias{create_ScanHistory_extended}
\title{create_ScanHistory_extended}
\usage{
create_ScanHistory_extended(chip_IDs, output_dir, result_ID)
}
\arguments{
\item{result_ID}{}
}
\value{

}
\description{
create_ScanHistory_extended
}
