% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImportBlobFiles.R,
%   R/helpers_MetadataScanPosition.R
\name{create_ScanHistory_of_chipIDs}
\alias{create_ScanHistory_of_chipIDs}
\title{create_ScanHistory_of_chipIDs}
\usage{
create_ScanHistory_of_chipIDs(chip_IDs)

create_ScanHistory_of_chipIDs(chip_IDs)
}
\arguments{
\item{chip_IDs}{}

\item{MethodHistory}{}
}
\value{



}
\description{
create_ScanHistory_of_chipIDs

create_ScanHistory_of_chipIDs
}
