% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_MetadataScanPosition.R
\name{create_ScanHistory_of_chipIDs}
\alias{create_ScanHistory_of_chipIDs}
\title{create_ScanHistory_of_chipIDs}
\usage{
create_ScanHistory_of_chipIDs(chip_IDs)
}
\arguments{
\item{chip_IDs}{}
}
\description{
create_ScanHistory_of_chipIDs
}
\examples{
\donttest{
data(chip_IDs)
SHs <- create_ScanHistory_of_chipIDs(chip_IDs)
}
}
\keyword{internal}
