% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImageProcessing.R
\name{export_image_result_tiffs}
\alias{export_image_result_tiffs}
\title{Title}
\usage{
export_image_result_tiffs(
  image,
  image_blurred,
  image_binary,
  output_dir,
  chip_ID,
  pos,
  sigma,
  threshold,
  cols,
  rows
)
}
\arguments{
\item{rows}{}
}
\value{

}
\description{
Title
}
