% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{extract_files_to_process}
\alias{extract_files_to_process}
\title{Extract the names of the files that are to be processed from an XML}
\usage{
extract_files_to_process(xml_dir)
}
\arguments{
\item{xml_dir}{Directory of the XML file}
}
\value{
Names of the files to be processed
}
\description{
Extract the names of the files that are to be processed from an XML
}
\seealso{
Other database related: 
\code{\link{connect_mongoDB}()},
\code{\link{extract_chip_IDs}()},
\code{\link{extract_gate_names}()},
\code{\link{extract_gates_to_process}()},
\code{\link{extract_jobtype}()},
\code{\link{find_server_path}()},
\code{\link{query_mongoDB}()}
}
\concept{database related}
\keyword{internal}
