% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{extract_gate_metadata}
\alias{extract_gate_metadata}
\title{Title}
\usage{
extract_gate_metadata(chip_IDs)
}
\arguments{
\item{chip_IDs}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
