% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{extract_jobtype}
\alias{extract_jobtype}
\title{Extract the job type that is to be executed}
\usage{
extract_jobtype(xml_dir)
}
\arguments{
\item{xml_dir}{Directory of the XML file}
}
\value{
Name of the job type
}
\description{
Extract the job type that is to be executed
}
\seealso{
Other database related: 
\code{\link{connect_mongoDB}()},
\code{\link{extract_chip_IDs}()},
\code{\link{extract_files_to_process}()},
\code{\link{extract_gate_names}()},
\code{\link{extract_gates_to_process}()},
\code{\link{find_all_attributes_in_EDL}()},
\code{\link{find_server_path}()},
\code{\link{find_valid_group_chip_IDs}()},
\code{\link{query_mongoDB}()}
}
\concept{database related}
\keyword{internal}
