% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_DBrelated.R
\name{find_chip_path}
\alias{find_chip_path}
\title{searches for a chip_ID}
\usage{
find_chip_path(ChannelID = "M583054", server_path = NULL)
}
\arguments{
\item{ChannelID}{character containing a chip_ID}

\item{server_path}{vector containing characters of server paths}
}
\value{
a character containing a chip_ID string or error message
}
\description{
finds the folder path to a related chip_ID in a list of server paths. If no serverpaths are provided to the function, a internal function is called.
}
\examples{
\donttest{
data(chip_IDs)
chip_path <- find_chip_path(chip_IDs[1])

}
}
\seealso{
Other database related: 
\code{\link{check_if_chip_data_exist}()},
\code{\link{connect_mongoDB}()},
\code{\link{extract_chipIDs_from_groupEDL}()},
\code{\link{find_server_path}()},
\code{\link{find_valid_group_chip_IDs}()},
\code{\link{get_EDL_from_query_result}()},
\code{\link{query_mongoDB}()}
}
\concept{database related}
\keyword{internal}
