% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_GenerateInputData.R
\name{find_server_path}
\alias{find_server_path}
\title{finds available server storing images generated by ZKWapp}
\usage{
find_server_path()
}
\description{
finds available server storing images generated by ZKWapp
}
\examples{
\donttest{
find_server_path()
}
}
\seealso{
Other database related: 
\code{\link{connect_mongoDB}()},
\code{\link{extract_chip_IDs}()},
\code{\link{extract_files_to_process}()},
\code{\link{extract_gate_names}()},
\code{\link{extract_gates_to_process}()},
\code{\link{extract_jobtype}()},
\code{\link{find_all_attributes_in_EDL}()},
\code{\link{query_mongoDB}()}
}
\concept{database related}
\keyword{internal}
