% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_DBrelated.R
\name{get_EDL_from_query_result}
\alias{get_EDL_from_query_result}
\title{Title}
\usage{
get_EDL_from_query_result(result)
}
\arguments{
\item{result}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
