% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_MetadataScanPosition.R
\name{get_sampleType_from_MethodHistory}
\alias{get_sampleType_from_MethodHistory}
\title{Title}
\usage{
get_sampleType_from_MethodHistory(MethodHistory)
}
\arguments{
\item{MethodHistory}{}
}
\description{
Title
}
\examples{
data("EDL_chipIDs_limslager")
EDL <- EDL_chipIDs_limslager[1]
MH <- create_MethodHistory_from_EDL(EDL)
sampleType <- get_sampleType_from_MethodHistory(MH)
}
\keyword{internal}
