% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImageProcessing.R
\name{perform_image_processing}
\alias{perform_image_processing}
\title{perform_image_processing}
\usage{
perform_image_processing(
  m.data,
  cellres = cellres,
  output_dir = output_dir,
  chip_ID,
  pos_ID,
  sigma = sigma,
  threshold = threshold
)
}
\arguments{
\item{threshold}{}
}
\value{

}
\description{
perform_image_processing
}
