% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImageProcessing.R
\name{process_TissueDetection}
\alias{process_TissueDetection}
\title{process_TissueDetection}
\usage{
process_TissueDetection(
  image_groups,
  output_dir,
  sigma = 15,
  threshold = 2,
  window = 10
)
}
\arguments{
\item{window}{}
}
\value{

}
\description{
process_TissueDetection
}
