% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ImageProcessing.R
\name{process_TissueDetection}
\alias{process_TissueDetection}
\title{process_TissueDetection}
\usage{
process_TissueDetection(
  image_groups,
  output_dir,
  sigma = 15,
  threshold = 2,
  window = 10,
  attenuation = 0.01,
  plot_image = TRUE,
  result_ID = ""
)
}
\arguments{
\item{attenuation}{}
}
\description{
process_TissueDetection
}
\examples{
output_dir <- "devel/data/data_output"
result_ID <- "P1761451"
}
