% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_DevelTools.R
\name{stop_if_fatal}
\alias{stop_if_fatal}
\title{Stop execution}
\usage{
stop_if_fatal(message = "Please see chunk output above.")
}
\arguments{
\item{message}{Message to be shown when execution is stopped}
}
\description{
Stop execution
}
\keyword{internal}
